% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsDPMalgo1.R
\name{gibbsDPMalgo1}
\alias{gibbsDPMalgo1}
\title{Gibbs Sampling with Alghorithm 1 from Neal}
\usage{
gibbsDPMalgo1(z, hyperG0, alpha, niter, doPlot = TRUE, pause = 0.1)
}
\arguments{
\item{z}{data matrix \code{p x n}}

\item{hyperG0}{a list of hyper parameters}

\item{alpha}{concentration parameter}

\item{niter}{number of MCMC iterations}

\item{doPlot}{whether to print plots. Default is \code{TRUE}.}

\item{pause}{how long to pause between graphical outputs. 
Necessary for proper display. Default is \code{0.1}s.}
}
\description{
Polya Urn representation
}
\examples{
rm(list=ls())
#Number of data
n <- 100
set.seed(1231)

# Sample data
m <- matrix(nrow=2, ncol=4, c(-1, 1, 0, 2, 1, -2, -1, -2))
p <- c(0.2, 0.1, 0.4, 0.3) # frequence des clusters

library(expm)
s <- array(dim=c(2,2,4))
s[, ,1] <- matrix(nrow=2, ncol=2, c(0.1, 0, 0, 0.1))
s[, ,2] <- matrix(nrow=2, ncol=2, c(0.01, 0, 0, 0.1))
s[, ,3] <- matrix(nrow=2, ncol=2, c(0.1, 0.08, 0.08, 0.1))
s[, ,4] <- .1*diag(2)
c <- rep(0,n)
z <- matrix(0, nrow=2, ncol=n)
for(k in 1:n){
 c[k] = which(rmultinom(n=1, size=1, prob=p)!=0)
 z[,k] <- m[, c[k]] + expm::sqrtm(s[, , c[k]])\%*\%matrix(rnorm(2, mean = 0, sd = 1), nrow=2, ncol=1)
}
 
 # Set parameters of G0
 hyperG0 <- list()
 hyperG0[["mu"]] <- c(0,0)
 hyperG0[["kappa"]] <- 1
 hyperG0[["nu"]] <- 4
 hyperG0[["lambda"]] <- diag(2)
 # Scale parameter of DPM
 alpha <- 3
 # Number of iterations
 N <- 30 
 # do some plots
 doPlot <- TRUE 
 
 plot(t(z), pch=16, xlim=c(-3,3), ylim=c(-3,3), xlab="Dim 1", ylab="Dim 2")
 # Gibbs sampler for Dirichlet Process Mixtures
 if(interactive()){
   GibSample <- gibbsDPMalgo1(z, hyperG0, alpha, niter = N, doPlot)
 gifski::save_gif(expr = gibbsDPMalgo1(z, hyperG0, alpha = 2, niter=15, doPlot=TRUE, pause = 0),
          delay = 0.02, width=600, height=500, gif_file = "BMgibbs_alpha2.gif")
 }


}
\author{
Francois Caron, Boris Hejblum
}
