% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsDPMalgo2.R
\name{gibbsDPMalgo2}
\alias{gibbsDPMalgo2}
\title{Gibbs Sampling with Alghorithm 2}
\usage{
gibbsDPMalgo2(z, hyperG0, alpha, niter, doPlot = TRUE, pause = 0.08)
}
\arguments{
\item{z}{data matrix \code{p x n}}

\item{hyperG0}{a list of hyper parameters}

\item{alpha}{concentration parameter}

\item{niter}{number of MCMC iterations}

\item{doPlot}{whether to print plots. Default is \code{TRUE}.}

\item{pause}{how long to pause between graphical outputs. 
Necessary for proper display. Default is \code{0.08}s.}
}
\description{
Chinese Restaurant Process representation
}
\examples{
rm(list=ls())
#Number of data
n <- 100
set.seed(1231)

# Sample data
m <- matrix(nrow=2, ncol=4, c(-1, 1, 0, 2, 1, -2, -1, -2))
p <- c(0.2, 0.1, 0.4, 0.3) # frequence des clusters

library(expm)
s <- array(dim=c(2,2,4))
s[, ,1] <- matrix(nrow=2, ncol=2, c(0.1, 0, 0, 0.1))
s[, ,2] <- matrix(nrow=2, ncol=2, c(0.01, 0, 0, 0.1))
s[, ,3] <- matrix(nrow=2, ncol=2, c(0.1, 0.08, 0.08, 0.1))
s[, ,4] <- .1*diag(2)
c <- rep(0,n)
z <- matrix(0, nrow=2, ncol=n)
for(k in 1:n){
 c[k] = which(rmultinom(n=1, size=1, prob=p)!=0)
 z[,k] <- m[, c[k]] + expm::sqrtm(s[, , c[k]])\%*\%matrix(rnorm(2, mean = 0, sd = 1), nrow=2, ncol=1)
}
 
 # Set parameters of G0
 hyperG0 <- list()
 hyperG0[["mu"]] <- c(0,0)
 hyperG0[["kappa"]] <- 1
 hyperG0[["nu"]] <- 4
 hyperG0[["lambda"]] <- diag(2)
 
 if(interactive()){
 # Gibbs sampler for Dirichlet Process Mixtures
 gifski::save_gif(expr = gibbsDPMalgo2(z, hyperG0, alpha = 0.0001, niter=15, 
                  doPlot=TRUE, pause = 0),
          delay = 0.02, width=600, height=500, gif_file = "CRPgibbs, alpha00001.gif")
 gifski::save_gif(expr = gibbsDPMalgo2(z, hyperG0, alpha = 2, niter=15, 
                  doPlot=TRUE, pause = 0),
          delay = 0.02, width=600, height=500, gif_file = "CRPgibbs, alpha2.gif")
 gifski::save_gif(expr = gibbsDPMalgo2(z, hyperG0, alpha = 30, niter=15, 
                  doPlot=TRUE, pause = 0),
          delay = 0.02, width=600, height=500, gif_file = "CRPgibbs, alpha30.gif")
          
          
 hyperG0[["lambda"]] <- 100*diag(2)
 gifski::save_gif(expr = gibbsDPMalgo2(z, hyperG0, alpha = 2, niter=15, 
                                       doPlot=TRUE, pause = 0),
          delay = 0.02, width=600, height=500, gif_file = "CRPgibbs, alpha2_lambda10.gif")
}

}
\author{
Francois Caron, Boris Hejblum
}
