% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{multivariate  Normal  probability density function}
\usage{
mvnpdf(x, mean, varcovM)
}
\arguments{
\item{x}{a p x n data matrix with n the number of observations and
p the number of dimensions}

\item{mean}{mean vector}

\item{varcovM}{variance-covariance matrix o}
}
\description{
multivariate  Normal  probability density function
}
