% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riW.R
\name{riW}
\alias{riW}
\title{Sample from a inverse-Wishart distribution}
\usage{
riW(n, lambda)
}
\arguments{
\item{n}{degrees of freedom}

\item{lambda}{scale parameter}
}
\description{
Sample from a inverse-Wishart distribution
}
