% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNiW.R
\name{rNiW}
\alias{rNiW}
\title{Sample from a Normal inverse-Wishart distribution
whose parameter are given by the structure hyper}
\usage{
rNiW(hyper, diagVar = TRUE)
}
\arguments{
\item{hyper}{hyper parameter list}

\item{diagVar}{whether to use a diagonal Covariance matrix.
Default is \code{TRUE}.}
}
\description{
Sample from a Normal inverse-Wishart distribution
whose parameter are given by the structure hyper
}
