% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rW.R
\name{rW}
\alias{rW}
\title{Sample from a Wishart distribution}
\usage{
rW(n, Sigma)
}
\arguments{
\item{n}{degrees of freedom}

\item{Sigma}{scale parameter (matrix)}
}
\description{
Sample from a Wishart distribution
}
