% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_SS.R
\name{update_SS}
\alias{update_SS}
\title{Update Sufficient Statistics}
\usage{
update_SS(z, S)
}
\arguments{
\item{z}{data matrix \code{p x n}}

\item{S}{list of NiW parameters}
}
\description{
Return updated sufficient statistics S of Normal Wishart distribution
with new data matrix z
}
